requirements.txt
# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pypdf>=3.17.0

# Knowledge Graph Generation
kg-gen>=0.1.0
litellm>=1.0.0
dspy-ai>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# OpenAI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Clustering and ML (for semantic model)
scikit-learn>=1.3.0
networkx>=3.0
matplotlib>=3.7.0
seaborn>=0.12.0
hdbscan>=0.8.33

# Graph visualization
pyvis>=0.3.2
